[Refactor] category apis
