+++ src/test/java/io/waterkite94/hd/hotdeal/item/web/api/admin/CategoryAdminControllerTest.java
@@ -1,4 +1,4 @@
-package io.waterkite94.hd.hotdeal.item.web.api;
+package io.waterkite94.hd.hotdeal.item.web.api.admin;
 
 import static org.mockito.BDDMockito.*;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.*;
@@ -9,8 +9,6 @@ import static org.springframework.security.test.web.servlet.request.SecurityMock
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
-import java.util.List;
-
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -20,51 +18,48 @@ import org.springframework.restdocs.payload.JsonFieldType;
 import org.springframework.security.test.context.support.WithMockUser;
 
 import io.waterkite94.hd.hotdeal.ControllerTestSupport;
-import io.waterkite94.hd.hotdeal.item.domain.Category;
 import io.waterkite94.hd.hotdeal.item.service.admin.CategoryService;
-import io.waterkite94.hd.hotdeal.item.web.api.normal.CategoryController;
+import io.waterkite94.hd.hotdeal.item.web.api.request.AddCategoryRequest;
 
-@WebMvcTest(CategoryController.class)
-class CategoryControllerTest extends ControllerTestSupport {
+@WebMvcTest(CategoryAdminController.class)
+class CategoryAdminControllerTest extends ControllerTestSupport {
 
 	@MockBean
 	private CategoryService categoryService;
 
 	@Test
 	@WithMockUser(value = "USER")
-	@DisplayName(value = "카테고리를 조회하는 API를 호출합니다.")
-	void findAllCategoriesApi() throws Exception {
+	@DisplayName(value = "카테고리를 추가하는 API를 호출합니다.")
+	void addCategory() throws Exception {
 		// given
-		Category category = createCategory();
-		given(categoryService.findAllCategories()).willReturn(List.of(category));
+		AddCategoryRequest request = createRequest();
+		given(categoryService.saveCategory(request.toServiceDto())).willReturn(1L);
 
 		// when // then
 		mockMvc.perform(
-				get("/api/v1/categories")
+				post("/api/v1/admin/category")
 					.with(csrf())
 					.contentType(MediaType.APPLICATION_JSON)
+					.content(objectMapper.writeValueAsString(request))
 			).andDo(print())
 			.andExpect(status().isOk())
-			.andExpect(jsonPath("$.status").isNumber())
-			.andExpect(jsonPath("$.message").isEmpty())
-			.andExpect(jsonPath("$.data").isArray())
-			.andExpect(jsonPath("$.data[0].id").value(category.getId()))
-			.andExpect(jsonPath("$.data[0].name").value(category.getName()))
-			.andDo(document("category-find-all",
+			.andExpect(jsonPath("$.data.category_id").isNumber())
+			.andDo(document("category-add",
 				preprocessRequest(prettyPrint()),
 				preprocessResponse(prettyPrint()),
+				requestFields(
+					fieldWithPath("name").type(JsonFieldType.STRING).description("카테고리 이름")
+				),
 				responseFields(
 					fieldWithPath("status").type(JsonFieldType.NUMBER).description("요청 상태 코드"),
 					fieldWithPath("message").type(JsonFieldType.NULL).description("요청 결과 메시지"),
-					fieldWithPath("data[].id").type(JsonFieldType.NUMBER).description("카테고리 아이디"),
-					fieldWithPath("data[].name").type(JsonFieldType.STRING).description("카테고리 이름")
+					fieldWithPath("data.category_id").type(JsonFieldType.NUMBER).description("카테고리 아이디")
 				)
 			));
 	}
 
-	private Category createCategory() {
-		return Category.builder()
-			.id(1L)
+	private AddCategoryRequest createRequest() {
+		return AddCategoryRequest.builder()
 			.name("name")
 			.build();
 	}
