+++ src/test/java/io/waterkite94/hd/hotdeal/item/web/api/normal/CategoryControllerTest.java
@@ -1,6 +1,5 @@
-package io.waterkite94.hd.hotdeal.item.web.api;
+package io.waterkite94.hd.hotdeal.item.web.api.normal;
 
-import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.BDDMockito.*;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.*;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.*;
@@ -9,6 +8,8 @@ import static org.springframework.restdocs.payload.PayloadDocumentation.*;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+
+import java.util.List;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -19,50 +20,50 @@ import org.springframework.restdocs.payload.JsonFieldType;
 import org.springframework.security.test.context.support.WithMockUser;
 
 import io.waterkite94.hd.hotdeal.ControllerTestSupport;
+import io.waterkite94.hd.hotdeal.item.domain.Category;
 import io.waterkite94.hd.hotdeal.item.service.admin.CategoryService;
-import io.waterkite94.hd.hotdeal.item.web.api.admin.CategoryAdminController;
-import io.waterkite94.hd.hotdeal.item.web.api.request.AddCategoryRequest;
 
-@WebMvcTest(CategoryAdminController.class)
-class CategoryAdminControllerTest extends ControllerTestSupport {
+@WebMvcTest(CategoryController.class)
+class CategoryControllerTest extends ControllerTestSupport {
 
 	@MockBean
 	private CategoryService categoryService;
 
 	@Test
 	@WithMockUser(value = "USER")
-	@DisplayName(value = "카테고리를 추가하는 API를 호출합니다.")
-	void addCategory() throws Exception {
+	@DisplayName(value = "카테고리를 조회하는 API를 호출합니다.")
+	void findAllCategoriesApi() throws Exception {
 		// given
-		AddCategoryRequest request = createRequest();
-
-		given(categoryService.saveCategory(any())).willReturn(1L);
+		Category category = createCategory();
+		given(categoryService.findAllCategories()).willReturn(List.of(category));
 
 		// when // then
 		mockMvc.perform(
-				post("/admin/category")
+				get("/api/v1/categories")
 					.with(csrf())
 					.contentType(MediaType.APPLICATION_JSON)
-					.content(objectMapper.writeValueAsString(request))
 			).andDo(print())
 			.andExpect(status().isOk())
-			.andExpect(jsonPath("$.data").isNumber())
-			.andDo(document("category-add",
+			.andExpect(jsonPath("$.status").isNumber())
+			.andExpect(jsonPath("$.message").isEmpty())
+			.andExpect(jsonPath("$.data").isArray())
+			.andExpect(jsonPath("$.data[0].id").value(category.getId()))
+			.andExpect(jsonPath("$.data[0].name").value(category.getName()))
+			.andDo(document("category-find-all",
 				preprocessRequest(prettyPrint()),
 				preprocessResponse(prettyPrint()),
-				requestFields(
-					fieldWithPath("name").type(JsonFieldType.STRING).description("카테고리 이름")
-				),
 				responseFields(
 					fieldWithPath("status").type(JsonFieldType.NUMBER).description("요청 상태 코드"),
 					fieldWithPath("message").type(JsonFieldType.NULL).description("요청 결과 메시지"),
-					fieldWithPath("data").type(JsonFieldType.NUMBER).description("카테고리 아이디")
+					fieldWithPath("data[].id").type(JsonFieldType.NUMBER).description("카테고리 아이디"),
+					fieldWithPath("data[].name").type(JsonFieldType.STRING).description("카테고리 이름")
 				)
 			));
 	}
 
-	private AddCategoryRequest createRequest() {
-		return AddCategoryRequest.builder()
+	private Category createCategory() {
+		return Category.builder()
+			.id(1L)
 			.name("name")
 			.build();
 	}
